Moved the PluginParameters and JavaToZodConvertorWrapper into the core package.
