+++ java-to-zod-core/pom.xml
@@ -49,5 +49,18 @@
       <version>3.9.2</version>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.apache.maven.plugin-tools</groupId>
+      <artifactId>maven-plugin-annotations</artifactId>
+      <version>3.10.2</version>
+      <scope>compile</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.gradle</groupId>
+      <artifactId>gradle-api</artifactId>
+      <version>8.5</version>
+      <scope>system</scope>
+      <systemPath>C:\Users\thele\.gradle\caches\8.5\generated-gradle-jars\gradle-api-8.5.jar</systemPath>
+    </dependency>
   </dependencies>
 </project>
